pyproject.toml
[project]
name = "ai-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "joblib>=1.5.2",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "panda>=0.3.1",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "streamlit>=1.51.0",
]
